requirements.txt
Flask==3.1.0
flask-cors==5.0.1
google-generativeai==0.8.5
gunicorn==20.1.0
